Redesign loggin interfaces
