+++ pkg/flow/states/generateEvent.go
@@ -8,8 +8,8 @@ import (
 
 	cloudevents "github.com/cloudevents/sdk-go/v2"
 	derrors "github.com/direktiv/direktiv/pkg/flow/errors"
-	log "github.com/direktiv/direktiv/pkg/flow/internallogger"
 	"github.com/direktiv/direktiv/pkg/model"
+	log "github.com/direktiv/direktiv/pkg/refactor/internallogger"
 	"github.com/google/uuid"
 	"github.com/senseyeio/duration"
 )
