+++ pkg/flow/states/switch.go
@@ -6,8 +6,8 @@ import (
 	"fmt"
 
 	derrors "github.com/direktiv/direktiv/pkg/flow/errors"
-	log "github.com/direktiv/direktiv/pkg/flow/internallogger"
 	"github.com/direktiv/direktiv/pkg/model"
+	log "github.com/direktiv/direktiv/pkg/refactor/internallogger"
 )
 
 func init() {
