+++ pkg/flow/states/validate.go
@@ -7,8 +7,8 @@ import (
 	"fmt"
 
 	derrors "github.com/direktiv/direktiv/pkg/flow/errors"
-	log "github.com/direktiv/direktiv/pkg/flow/internallogger"
 	"github.com/direktiv/direktiv/pkg/model"
+	log "github.com/direktiv/direktiv/pkg/refactor/internallogger"
 	"github.com/xeipuuv/gojsonschema"
 )
