+++ pkg/refactor/internallogger/mockdata/entlog_loop.json
