+++ pkg/refactor/internallogger/mockdata/entlog_loopFunctionNested.json
