+++ pkg/refactor/utils/utils.go
@@ -1,4 +1,4 @@
-package testutils
+package utils
 
 import (
 	"context"
